project1/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.10.0
gunicorn>=21.2.0
